Split ApplicationController #14
